==> 02b_isoform_analysis_173955.qmd <==
---
title: "02_isoform_analysis"
format: html
---

```{r}
#| message: false
#| warning: false
library(DEXSeq)
library(fgsea)
library(msigdbr)
library(tidyverse)
library(tidyr)
```

```{r}
library(DEXSeq)

# keep only isoforms with >=10 counts in 0.5 of samples
min_samples <- ncol(GSE173955_isoform) * 0.5

# Filter isoforms that have >= 10 counts in at least 50% of samples
filtered_isoforms <- GSE173955_isoform[
  rowSums(GSE173955_isoform >= 10) >= min_samples,
]

# Print the dimensions to see how many isoforms were kept
cat("Original dimensions:", dim(GSE173955_isoform), "\n")
cat("Filtered dimensions:", dim(filtered_isoforms), "\n")
cat("Percentage of isoforms retained:", 
    round(nrow(filtered_isoforms)/nrow(GSE173955_isoform)*100, 2), "%\n")

# Ensure metadata rows are in same order as count columns
GSE173955_metadata <- GSE173955_metadata[
    match(colnames(GSE173955_isoform), GSE173955_metadata$sample_id),
]

all(colnames(GSE173955_isoform) == GSE173955_metadata$sample_id)

```

```{r}
head(rownames(GSE173955_isoform))
str(rownames(GSE173955_isoform))


feature_ids <- as.character(rownames(GSE173955_isoform))
```

```{r}



isoInfo <- read_rds(
  paste0(
  '/home/projects/22140/exam_data/',
  'Homo_sapiens.GRCh38.107.chr_patch_hapl_scaff_tx_info.Rds'
  )
)

isoInfo <- isoInfo[match(rownames(filtered_isoforms), isoInfo$transcript_id), ]

# Get the valid sample names from metadata
valid_samples <- GSE173955_metadata$sample_id

# Keep only columns whose names appear in metadata
filtered_isoforms <- filtered_isoforms[, colnames(filtered_isoforms) %in% valid_samples]
GSE173955_metadata$gender <- factor(GSE173955_metadata$conditions)

GSE173955_metadata = na.omit(GSE173955_metadata)


dxd <- DEXSeqDataSet(
    countData = filtered_isoforms,
    sampleData = GSE173955_metadata,
    design = ~ sample + exon + conditions:exon + gender:exon,
    featureID = isoInfo$transcript_id,
    groupID = isoInfo$gene_id
)

```

```{r}
dxd <- estimateSizeFactors(dxd)
dxd <- estimateDispersions(dxd)

```

```{r}
library(DEXSeq)

dxd <- testForDEU(dxd)
dxd <- estimateExonFoldChanges(dxd, fitExpToVar = "conditions")

res <- DEXSeqResults(dxd)

GSE173955_isoform_results = res

save(GSE173955_isoform_results, 
     file = "/home/projects/exam_2025_22140_22150/group_2/data/GSE173955_isoform_results.Rdata")

```

```{r}
alpha <- 0.05
# Significant isoforms (rows)
sig_iso <- res[!is.na(res$padj) & res$padj < alpha, ]
n_sig_isoforms <- nrow(sig_iso)
n_sig_isoforms

n_sig_genes <- length(unique(sig_iso$groupID))
n_sig_genes
```

```{r}
dexseq_df <- as.data.frame(res)
dexseq_df <- dexseq_df[order(dexseq_df$padj), ]
top5_genes <- dexseq_df$groupID[!is.na(dexseq_df$padj)][1:6]

top5_genes
```
